DBE - UI + Business

+++ C#/Project/PWC Bilingual Publication/EuCA.Pwc.Pub/PublicationProcess.cs
@@ -1,16 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using log4net;
+using System;
+using System.Configuration;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 
 namespace EuCA.Pwc.Pub
 {
     public class PublicationManager
     {
+        #region Members
+
+        private static readonly ILog _logger = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+
+        #endregion
+
         #region Singleton
 
         // static holder for instance, need to use lambda to construct since constructor private
@@ -36,31 +41,46 @@ namespace EuCA.Pwc.Pub
 
             // 1. Generate source (english) HTML files
             progress.Report("1. Generate source (english) HTML files");
-            //GenerateSourceHtml(parameters, progress, cancellationToken);
+            _logger.Debug("Generating source (english) HTML files");
+            GenerateSourceHtml(parameters, progress, cancellationToken);
 
             // 2. Generate translated HTML files
             progress.Report("2. Generate translated HTML files");
-            //GenerateTranslationHtml(parameters, progress, cancellationToken);
+            _logger.Debug("Generating translated HTML files");
+            GenerateTranslationHtml(parameters, progress, cancellationToken);
 
             // 3. Merge HTML files
             progress.Report("3. Merge HTML files");
+            _logger.Debug("Merging HTML files");
             MergeHtmlFiles(parameters, progress, cancellationToken);
 
             // 4.a. Process meta data for banner
             progress.Report("4.a. Process meta data for banner");
+            _logger.Debug("Processing meta data for banner");
             ProcessMetadata(parameters, progress, cancellationToken);
 
             // 4.b. Process TOC
             progress.Report("4.b. Process TOC");
+            _logger.Debug("Processing TOC");
             ProcessToc(parameters, progress, cancellationToken);
 
             // 5. Prepare package
             progress.Report("5. Prepare package");
+            _logger.Debug("Preparing package");
             PreparePackage(parameters, progress, cancellationToken);
 
-            // 6. Delete temporary files and directories
-            progress.Report("6. Delete temporary files and directories");
-            DeleteTemp(parameters, progress, cancellationToken);
+            // 6. Copy graphics
+            progress.Report("6. Copy graphics");
+            _logger.Debug("Copying graphics");
+            CopyGraphics(parameters, progress, cancellationToken);
+
+            // 7. Delete temporary files and directories
+            progress.Report("7. Delete temporary files and directories");
+            if (parameters.DeleteTempFile)
+            {
+                _logger.Debug("Deleting temporary files and directory");
+                DeleteTemp(parameters, progress, cancellationToken);
+            }
 
             progress.Report("Publicaton process end.");
         }
@@ -72,11 +92,13 @@ namespace EuCA.Pwc.Pub
             var cmd =
                 "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform " +
                 "-s:\"" + p.FileOrig + "\" " +
-                "-xsl:\"" + p.XslDir + "\\XSL\\jmtosmigrate.xsl\" " +
+                "-xsl:\"" + p.DirXsl + "\\jmtosmigrate.xsl\" " +
                 "-o:" + Path.GetFileNameWithoutExtension(p.FileOrig) + ".html " +
                 "output-dir=\"" + new Uri(p.DirOrig).AbsoluteUri  + "\"";
 
             RunProcess(cmd);
+
+            File.Delete(Path.Combine(Environment.CurrentDirectory, Path.GetFileNameWithoutExtension(p.FileOrig) + ".html"));
         }
 
         private void GenerateTranslationHtml(ProcessParameters p, IProgress<object> progress, CancellationToken cancellationToken)
@@ -86,11 +108,13 @@ namespace EuCA.Pwc.Pub
             var cmd =
                "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform " +
                "-s:\"" + p.FileTrad + "\" " +
-               "-xsl:\"" + p.XslDir + "\\XSL\\jmtosmigrate.xsl\" " +
+               "-xsl:\"" + p.DirXsl + "\\jmtosmigrate.xsl\" " +
                "-o:" + Path.GetFileNameWithoutExtension(p.FileTrad) + ".html " +
                "output-dir=\"" + new Uri(p.DirTrad).AbsoluteUri + "\"";
 
             RunProcess(cmd);
+
+            File.Delete(Path.Combine(Environment.CurrentDirectory, Path.GetFileNameWithoutExtension(p.FileTrad) + ".html"));
         }
 
         private void MergeHtmlFiles(ProcessParameters p, IProgress<object> progress, CancellationToken cancellationToken)
@@ -111,47 +135,54 @@ namespace EuCA.Pwc.Pub
                 var fileIds = Path.Combine(p.DirTradIds, fileName); // Output file path of the transformation
                 if (!File.Exists(Path.Combine(p.DirTradPages, fileName))) // All the file that do no exist in the translated data (no need to merge)
                 {
+                    _logger.Debug(Path.GetFileNameWithoutExtension(file) + ": english only");
                     // Copy the file to the IDS directory
                     File.Copy(file, fileIds, true);
 
-                    RemoveNamespaces(fileIds, p.XslDir, p.DirTrad, cancellationToken);
+                    RemoveNamespaces(fileIds, p.DirXslBil, p.DirTrad, cancellationToken);
 
                     // "Title page" processing
                     var cmd =
                       "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform" +
                       " -s:\"" + fileIds + "\"" +
-                      " -xsl:\"" + p.XslDir + "\\XSL_BIL\\bil_title-page.xsl\"" +
+                      " -xsl:\"" + p.DirXslBil + "\\bil_title-page.xsl\"" +
                       " -o:" + Path.Combine(p.DirTradPackageContent, fileName);
                     RunProcess(cmd);
                 }
                 else // File exists for english and chinese
                 {
+                    _logger.Debug(Path.GetFileNameWithoutExtension(file) + ": both language");
+
                     var fileTradIds = Path.Combine(p.DirTradIds, Path.GetFileNameWithoutExtension(fileName) + ".zh.html");
                     var fileOrigIds = Path.Combine(p.DirTradIds, Path.GetFileNameWithoutExtension(fileName) + ".en.html");
 
                     // Copy translated files and add language to the filename
                     File.Copy(Path.Combine(p.DirTradPages, fileName), fileTradIds, true);
 
-                    RemoveNamespaces(fileTradIds, p.XslDir, p.DirTrad, cancellationToken);
+                    RemoveNamespaces(fileTradIds, p.DirXslBil, p.DirTrad, cancellationToken);
                     // Copy original files and add language to the filename
                     File.Copy(Path.Combine(p.DirOrigPages, fileName), fileOrigIds, true);
 
-                    RemoveNamespaces(fileOrigIds, p.XslDir, p.DirTrad, cancellationToken);
+                    RemoveNamespaces(fileOrigIds, p.DirXslBil, p.DirTrad, cancellationToken);
+
+                    _logger.Debug(Path.GetFileNameWithoutExtension(file) + ": set ids");
 
                     // Set the ids of the translated data
                     var cmd =
                        "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform -dtd:off" +
                        " -s:\"" + fileTradIds + "\"" +
-                       " -xsl:\"" + p.XslDir + "\\XSL_BIL\\bil_set_ids.xsl\"" +
+                       " -xsl:\"" + p.DirXslBil + "\\bil_set_ids.xsl\"" +
                        " -o:" + fileIds +
                        " src-doc-path=" + fileOrigIds;
                     RunProcess(cmd);
+
+                    _logger.Debug(Path.GetFileNameWithoutExtension(file) + ": merging");
 
                     // Display content as tables
                     cmd =
                       "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform" +
                       " -s:\"" + fileIds + "\"" +
-                      " -xsl:\"" + p.XslDir + "\\XSL_BIL\\bil_merge.xsl\"" +
+                      " -xsl:\"" + p.DirXslBil + "\\bil_merge.xsl\"" +
                       " -o:" + Path.Combine(p.DirTradMerge, fileName);
                     RunProcess(cmd);
                 }
@@ -163,12 +194,12 @@ namespace EuCA.Pwc.Pub
             cancellationToken.ThrowIfCancellationRequested();
 
             var fileMeta = Path.Combine(p.DirOrigXml, "Metadata.xml");
-            RemoveNamespaces(fileMeta, p.XslDir, p.DirTrad, cancellationToken);
+            RemoveNamespaces(fileMeta, p.DirXslBil, p.DirTrad, cancellationToken);
 
             var cmd =
                 "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform" +
                 " -s:\"" + fileMeta + "\"" +
-                " -xsl:\"" + p.XslDir + "\\XSL_BIL\\bil_mtdt.xsl\"" +
+                " -xsl:\"" + p.DirXslBil + "\\bil_mtdt.xsl\"" +
                 " -o:" + Path.Combine(p.DirTradPackageContent, "metadata.html");
             RunProcess(cmd);
         }
@@ -181,7 +212,7 @@ namespace EuCA.Pwc.Pub
             var cmd =
                 "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform" +
                 " -s:\"" + fileToc + "\"" +
-                " -xsl:\"" + p.XslDir + "\\XSL_BIL\\bil_toc.xsl\"" +
+                " -xsl:\"" + p.DirXslBil + "\\bil_toc.xsl\"" +
                 " -o:" + Path.Combine(p.DirTradPackageContent, "index.html");
             RunProcess(cmd);
         }
@@ -190,7 +221,7 @@ namespace EuCA.Pwc.Pub
         {
             cancellationToken.ThrowIfCancellationRequested();
 
-            DeepCopy(new DirectoryInfo(Path.Combine(p.XslDir, "Lib_BIL")), new DirectoryInfo(p.DirTradPackageContent));
+            DeepCopy(new DirectoryInfo(Path.Combine(Environment.CurrentDirectory, ConfigurationManager.AppSettings["pwc_lib_bil"])), new DirectoryInfo(p.DirTradPackageContent));
             if (File.Exists(Path.Combine(p.DirTradPackage, "launch.exe"))) File.Delete(Path.Combine(p.DirTradPackage, "launch.exe"));
             File.Move(Path.Combine(p.DirTradPackageContent, "launch.exe"), Path.Combine(p.DirTradPackage, "launch.exe"));
 
@@ -206,14 +237,40 @@ namespace EuCA.Pwc.Pub
             source.EnumerateFiles().ToList().ForEach(file => file.CopyTo(Path.Combine(target.FullName, file.Name), true));
         }
 
+        private void CopyGraphics(ProcessParameters p, IProgress<object> progress, CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            var dirGraphics = Path.Combine(Directory.GetParent(p.FileOrig).FullName, "graphics");
+            if (!Directory.Exists(dirGraphics)) return;
+
+            Directory.CreateDirectory(p.DirGraphics);
+
+            Directory
+                .EnumerateFiles(dirGraphics, "*.*", SearchOption.AllDirectories)
+                .ToList()
+                .ForEach(file => {
+                    _logger.Debug(Path.GetFileName(file) + ": copying");
+                    File.Copy(file, Path.Combine(p.DirGraphics, Path.GetFileName(file)), true);
+                });
+        }
+
         private void DeleteTemp(ProcessParameters p, IProgress<object> progress, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
+
+            try
+            {
+                Directory.Delete(p.DirTrad, true);
+                Directory.Delete(p.DirOrig, true);
+            } 
+            catch (Exception) { }
         }
 
         private void RemoveNamespaces(string file, string dirXsl, string dirTmp, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
+
+            _logger.Debug(Path.GetFileNameWithoutExtension(file) + ": removing namespace");
 
             var fileTmp = Path.Combine(dirTmp, "tmp.xml");
 
@@ -225,97 +282,12 @@ namespace EuCA.Pwc.Pub
             var cmd =
                "/c java -cp \"./lib/java/saxonb9-1-0-8j/saxon9.jar\" net.sf.saxon.Transform" +
                " -s:\"" + file + "\"" +
-               " -xsl:\"" + dirXsl + "\\XSL_BIL\\bil_del_ns.xsl\"" +
+               " -xsl:\"" + dirXsl + "\\bil_del_ns.xsl\"" +
                " -o:" + fileTmp;
             RunProcess(cmd);
             File.Copy(fileTmp, file, true);
             File.Delete(fileTmp);
         }
-
-        /*private async void XSLTransform(string cmd, IProgress<object> progress, CancellationToken cancellationToken)
-        {
-            var t = RunProcessAsync(cmd, cancellationToken);
-
-            await t.ContinueWith((task) =>
-            {
-                if (task.Exception != null)
-                {
-                    // Throw exception
-                }
-                else if (task.IsCanceled)
-                {
-
-                }
-                else
-                {
-
-                }
-            });
-        }
-        
-        private Task<ProcessResults> RunProcessAsync(string command, CancellationToken cancellationToken)
-        {
-            // Creates a new ProcessStartInfo with the command
-            var processStartInfo = new ProcessStartInfo("cmd.exe", "/c " + command);
-
-            // Forces some settings in the start info so we can capture the output
-            processStartInfo.UseShellExecute = false;
-            processStartInfo.RedirectStandardOutput = true;
-            processStartInfo.RedirectStandardError = true;
-            
-            // Creates an instance of a TaskCompletionSource
-            var tcs = new TaskCompletionSource<ProcessResults>();
-
-            // List that will hold the outputs of the process
-            var standardOutput = new List<string>();
-            var standardError = new List<string>();
-
-            // Create a new process
-            var process = new Process
-            {
-                StartInfo = processStartInfo, // Process Start Info
-                EnableRaisingEvents = true // Enable proces to raise events (allow us to read the outputs)
-            };
-
-            // Handles the exit of the process
-            process.Exited += (sender, args) =>
-            {
-                // Handles error
-                if (process.ExitCode != 0) tcs.TrySetException(new InvalidOperationException(string.Join("\n", standardError.ToArray())));
-
-                // Set the result
-                tcs.TrySetResult(new ProcessResults(process, standardOutput, standardError));
-
-                // Dispose the process
-                process.Dispose();
-            };
-
-            // Handles the output data
-            process.OutputDataReceived  +=  (sender, args)  =>    {  if (args.Data != null) standardOutput.Add(args.Data);    };
-            // Handle the error data
-            process.ErrorDataReceived   +=  (sender, args)  =>    {  if (args.Data != null) standardError.Add(args.Data);     };
-
-            // Handles the cancellation of the task
-            cancellationToken.Register(() =>
-            {
-                // Try to cancel the task
-                tcs.TrySetCanceled();
-                // Dispose the process
-                process.Dispose();
-            });
-
-            cancellationToken.ThrowIfCancellationRequested();
-
-            // Start process
-            if (process.Start() == false)  tcs.TrySetException(new InvalidOperationException("Failed to start process"));
-
-            // Begin reading the outputs
-            process.BeginOutputReadLine();
-            process.BeginErrorReadLine();
-
-            // Return the task
-            return tcs.Task;
-        }*/
 
         private string RunProcess(string command, string fileName = "cmd.exe") 
         {
@@ -325,6 +297,8 @@ namespace EuCA.Pwc.Pub
             processStartInfo.UseShellExecute = false;
             processStartInfo.RedirectStandardOutput = true;
             processStartInfo.RedirectStandardError = true;
+            processStartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            processStartInfo.CreateNoWindow = true;
             // Create a new process
             var process = new Process
             {
@@ -343,33 +317,4 @@ namespace EuCA.Pwc.Pub
 
         #endregion
     }
-    /*
-    public class ProcessResults
-    {
-        private readonly Process _process;
-        private readonly IEnumerable<string> _standardOutput;
-        private readonly IEnumerable<string> _standardError;
-
-        public ProcessResults(Process process, IEnumerable<string> standardOutput, IEnumerable<string> standardError)
-        {
-            _process = process;
-            _standardOutput = standardOutput;
-            _standardError = standardError;
-        }
-
-        public Process Process
-        {
-            get { return _process; }
-        }
-
-        public IEnumerable<string> StandardOutput
-        {
-            get { return _standardOutput; }
-        }
-
-        public IEnumerable<string> StandardError
-        {
-            get { return _standardError; }
-        }
-    }*/
 }
